Fixe wrong temperature handling. (#952)
* Fixed wrong temperature handling.

* Added static cast to uint8_t.

+++ system/system_monitor/include/hdd_reader/hdd_reader.h
@@ -35,7 +35,8 @@ struct HDDInfo
   int error_code_;      //!< @brief error code, 0 on success, otherwise error
   std::string model_;   //!< @brief Model number
   std::string serial_;  //!< @brief Serial number
-  int temp_;            //!< @brief temperature(DegC)
+  uint8_t temp_;        //!< @brief temperature(DegC)
+    // Lowest byte of the raw value contains the exact temperature value (Celsius degrees) in S.M.A.R.T. information.
 
   /**
    * @brief Load or save data members.
